chore: move helpers to differente file
